Add Rresult msg testable
This can be used (for example) as `result int rresult_msg` to test
for an integer success result with an Rresult.R.msg error case.

The Rresult dependency here could be removed, but it is nice to
be able to use the `Rresult.pp_msg` function directly.  The extra
dependency would only kick in for test cases in dependent users
of Alcotest.