Show one failure when multiple tests fail
This commit changes the behavior of Alcotest, so that without
--show-errors and --verbose, if there are any failures, the first
failure is printed to standard output.

In current Alcotest, if one test fails, Alcotest shows the failure.
However, if multiple tests fail, they are written to some log, which is
difficult to use. Standard output receives only a list of names of the
failed tests.

Also in current Alcotest, the --show-errors and --verbose flags cause
all the failures to be printed to standard output. However, this can be
overwhelming if a large number of tests fail. Worse, in a progressive
test suite, the first failure is the most relevant, so it would be nice
if it was printed last. Then, the interactive user wouldn't have to
scroll up through the entire failure list to find it. However, Alcotest
prints the failures in order, so it is the last, and least relevant,
failure which appears last.