improve help text and error handling for filtering
Use Cmdliner for converting int to string, and validating the number of
arguments.
Previously just a Failure message was shown with no clear indication to
what is wrong, now more details are shown:

```diff
--- old
+++ new
@@ -1,15 +1,19 @@
 + ./simple test test_set invalidnumber
+simple: TESTCASE argument: invalid value `invalidnumber', expected an integer
+Usage: simple test [OPTION]... [NAME] [TESTCASE]
+Try `simple test --help' or `simple --help' for more information.
 Testing My first test.
 + echo '$?=' 1
 $?= 1

 + ./simple test nonexistent_name
+Invalid request (no tests to run, filter skipped everything)!
 Testing My first test.
 + echo '$?=' 1
 $?= 1

 + ./simple test test_set 1 trailingarguments
-simple: internal error, uncaught exception:
-        Failure("filter_test")
-
+simple: too many arguments, don't know what to do with `trailingarguments'
+Usage: simple test [OPTION]... [NAME] [TESTCASE]
+Try `simple test --help' or `simple --help' for more information.
 Testing My first test.
 + echo '$?=' 1
 $?= 1
```

This does not change the semantics of what CLI arguments are accepted,
just the help text and error reporting.
From the usage example it is clear now that only 2 filters can be used.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>