Be able to allocate a formatter for stdout/stderr and use it instead of using OCaml's formatter - fix some data-race when some parallel tasks use the same formatter.
This commit tries to solve an issue about OCaml 5 and parallelism.
Alcotest uses OCaml's formatter which can be used by some libraries such
as Logs. In this situation and if Logs is used into a parallel task, a
data-race exists about the internal queue used by formatters.

This commit allows the user to allocate its own formatter for
stdout/stderr and give them to Alcotest.